draft(v2.0.0): Proposal for v2.0.0
- addition: adds `header` property to schema
- change: hides less useful/common `types` mentioned in `conventionalcommits@1.0.0-beta.4`

- restructures tests to be better suited for multiple versions
