+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "pretest": "standard",
-    "test": "npx mocha test.js"
+    "test": "npx mocha tests/*.test.js"
   },
   "repository": {
     "type": "git",
