+++ tests/v1.test.js
@@ -1,0 +1,7 @@
+/* global describe */
+const schema = require('../versions/1.0.0/schema.json')
+const shared = require('./shared')
+
+describe('v1.0.0', function () {
+  shared(schema)
+})
